fixed paths
made a big no no and updated all paths to persistent paths, little did i know that github pages and local server has an additional path added to the end of the url so that basically broke everything on the actual site. Everything has been reverted to relative paths as annoying as it is its better for me as a developer to still be able to use live server for testing.